poster dan durasi di lowongan

+++ sirase/database/migrations/2025_11_07_140824_create_lowongan_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->text('deskripsi');
             $table->text('kualifikasi');
             $table->string('posisiLowongan');
-            $table->unsignedInteger('durasiKerja');
+            $table->double('durasiKerja');
             $table->date('awalPendaftaran');
             $table->date('batasPendaftaran');
             $table->date('mulaiKerja');
